# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Marketing Agent API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Business {
  id: number;
  name: string;
  industry: string;
  description: string;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  targetAudience?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface BusinessInput {
  name: string;
  industry: string;
  description: string;
  website?: string;
  targetAudience?: string;
}

export interface BusinessUpdate {
  name?: string;
  industry?: string;
  description?: string;
  website?: string;
  targetAudience?: string;
}

export type BusinessStatsContentByType = {[key: string]: number};

export interface BusinessStats {
  totalContent: number;
  activeCampaigns: number;
  totalKeywords: number;
  totalCompetitors: number;
  contentByType: BusinessStatsContentByType;
}

export type ContentPieceType = typeof ContentPieceType[keyof typeof ContentPieceType];


export const ContentPieceType = {
  blog: 'blog',
  social: 'social',
  email: 'email',
  ad: 'ad',
  strategy: 'strategy',
} as const;

export type ContentPieceStatus = typeof ContentPieceStatus[keyof typeof ContentPieceStatus];


export const ContentPieceStatus = {
  draft: 'draft',
  published: 'published',
  archived: 'archived',
} as const;

export interface ContentPiece {
  id: number;
  businessId: number;
  type: ContentPieceType;
  title: string;
  body: string;
  /** @nullable */
  platform?: string | null;
  /** @nullable */
  tone?: string | null;
  status: ContentPieceStatus;
  createdAt: string;
}

export type ContentRequestType = typeof ContentRequestType[keyof typeof ContentRequestType];


export const ContentRequestType = {
  blog: 'blog',
  social: 'social',
  email: 'email',
  ad: 'ad',
  strategy: 'strategy',
} as const;

export interface ContentRequest {
  type: ContentRequestType;
  topic: string;
  tone?: string;
  platform?: string;
  additionalContext?: string;
}

export type CampaignStatus = typeof CampaignStatus[keyof typeof CampaignStatus];


export const CampaignStatus = {
  draft: 'draft',
  active: 'active',
  paused: 'paused',
  completed: 'completed',
} as const;

export type CampaignChannel = typeof CampaignChannel[keyof typeof CampaignChannel];


export const CampaignChannel = {
  social: 'social',
  email: 'email',
  seo: 'seo',
  ads: 'ads',
  content: 'content',
  multi: 'multi',
} as const;

export interface Campaign {
  id: number;
  businessId: number;
  name: string;
  description: string;
  status: CampaignStatus;
  channel: CampaignChannel;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  endDate?: string | null;
  /** @nullable */
  budget?: number | null;
  createdAt: string;
}

export type CampaignInputChannel = typeof CampaignInputChannel[keyof typeof CampaignInputChannel];


export const CampaignInputChannel = {
  social: 'social',
  email: 'email',
  seo: 'seo',
  ads: 'ads',
  content: 'content',
  multi: 'multi',
} as const;

export interface CampaignInput {
  name: string;
  description: string;
  channel: CampaignInputChannel;
  startDate?: string;
  endDate?: string;
  budget?: number;
}

export type CampaignUpdateStatus = typeof CampaignUpdateStatus[keyof typeof CampaignUpdateStatus];


export const CampaignUpdateStatus = {
  draft: 'draft',
  active: 'active',
  paused: 'paused',
  completed: 'completed',
} as const;

export type CampaignUpdateChannel = typeof CampaignUpdateChannel[keyof typeof CampaignUpdateChannel];


export const CampaignUpdateChannel = {
  social: 'social',
  email: 'email',
  seo: 'seo',
  ads: 'ads',
  content: 'content',
  multi: 'multi',
} as const;

export interface CampaignUpdate {
  name?: string;
  description?: string;
  status?: CampaignUpdateStatus;
  channel?: CampaignUpdateChannel;
  startDate?: string;
  endDate?: string;
  budget?: number;
}

export interface CampaignsSummary {
  draft: number;
  active: number;
  paused: number;
  completed: number;
  total: number;
}

export interface ChartDataPoint {
  label: string;
  value: number;
}

export interface Analytics {
  period: string;
  contentGenerated: number;
  campaignsRun: number;
  keywordsTracked: number;
  competitorsAnalyzed: number;
  contentByType: ChartDataPoint[];
  campaignActivity: ChartDataPoint[];
}

export interface ActivityItem {
  id: number;
  type: string;
  description: string;
  createdAt: string;
}

export interface Keyword {
  id: number;
  businessId: number;
  keyword: string;
  /** @nullable */
  searchVolume?: number | null;
  /** @nullable */
  difficulty?: number | null;
  /** @nullable */
  intent?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface SeoResearchRequest {
  topic: string;
  additionalContext?: string;
}

export interface Competitor {
  id: number;
  businessId: number;
  name: string;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  strengths?: string | null;
  /** @nullable */
  weaknesses?: string | null;
  /** @nullable */
  lastAnalyzedAt?: string | null;
  createdAt: string;
}

export interface CompetitorInput {
  name: string;
  website?: string;
  notes?: string;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiConversationInput {
  title: string;
}

export interface OpenaiMessageInput {
  content: string;
}

export interface OpenaiError {
  error: string;
}