# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Marketing Agent API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /businesses:
    get:
      operationId: listBusinesses
      tags: [businesses]
      summary: List all businesses
      responses:
        "200":
          description: List of businesses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Business"
    post:
      operationId: createBusiness
      tags: [businesses]
      summary: Create a new business
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessInput"
      responses:
        "201":
          description: Created business
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
  /businesses/{id}:
    get:
      operationId: getBusiness
      tags: [businesses]
      summary: Get a business by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Business details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
        "404":
          description: Not found
    patch:
      operationId: updateBusiness
      tags: [businesses]
      summary: Update a business
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessUpdate"
      responses:
        "200":
          description: Updated business
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
    delete:
      operationId: deleteBusiness
      tags: [businesses]
      summary: Delete a business
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /businesses/{id}/stats:
    get:
      operationId: getBusinessStats
      tags: [businesses]
      summary: Get aggregated marketing stats for a business
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Business marketing stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessStats"

  # --- Content Generation ---
...
  /businesses/{businessId}/campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List marketing campaigns for a business
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a campaign
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignInput"
      responses:
        "201":
          description: Created campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /businesses/{businessId}/campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get a campaign
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    patch:
      operationId: updateCampaign
      tags: [campaigns]
      summary: Update a campaign
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignUpdate"
      responses:
        "200":
          description: Updated campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    delete:
      operationId: deleteCampaign
      tags: [campaigns]
      summary: Delete a campaign
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /businesses/{businessId}/campaigns/summary:
    get:
      operationId: getCampaignsSummary
      tags: [campaigns]
      summary: Get campaign status breakdown for a business
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign summary by status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignsSummary"

  # --- Analytics ---
  /businesses/{businessId}/analytics:
    get:
      operationId: fetchAnalytics
      tags: [analytics]
      summary: Get analytics data for a business
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Analytics data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Analytics"
  /businesses/{businessId}/analytics/activity:
    get:
      operationId: getActivityFeed
      tags: [analytics]
      summary: Get recent marketing activity for a business
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Activity feed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"

  # --- SEO ---
...
components:
  schemas:
...
    Business:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        industry:
          type: string
        description:
          type: string
        website:
          type: ["string", "null"]
        targetAudience:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - industry
        - description
        - createdAt
        - updatedAt
    BusinessInput:
      type: object
      properties:
        name:
          type: string
        industry:
          type: string
        description:
          type: string
        website:
          type: string
        targetAudience:
          type: string
      required:
        - name
        - industry
        - description
    BusinessUpdate:
      type: object
      properties:
        name:
          type: string
        industry:
          type: string
        description:
          type: string
        website:
          type: string
        targetAudience:
          type: string
    BusinessStats:
      type: object
      properties:
        totalContent:
          type: integer
        activeCampaigns:
          type: integer
        totalKeywords:
          type: integer
        totalCompetitors:
          type: integer
        contentByType:
          type: object
          additionalProperties:
            type: integer
      required:
        - totalContent
        - activeCampaigns
        - totalKeywords
        - totalCompetitors
        - contentByType

    # --- Content schemas ---
...
    Campaign:
      type: object
      properties:
        id:
          type: integer
        businessId:
          type: integer
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [draft, active, paused, completed]
        channel:
          type: string
          enum: [social, email, seo, ads, content, multi]
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        budget:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - businessId
        - name
        - description
        - status
        - channel
        - createdAt
    CampaignInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        channel:
          type: string
          enum: [social, email, seo, ads, content, multi]
        startDate:
          type: string
        endDate:
          type: string
        budget:
          type: number
      required:
        - name
        - description
        - channel
    CampaignUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [draft, active, paused, completed]
        channel:
          type: string
          enum: [social, email, seo, ads, content, multi]
        startDate:
          type: string
        endDate:
          type: string
        budget:
          type: number
    CampaignsSummary:
      type: object
      properties:
        draft:
          type: integer
        active:
          type: integer
        paused:
          type: integer
        completed:
          type: integer
        total:
          type: integer
      required:
        - draft
        - active
        - paused
        - completed
        - total

    # --- Analytics schemas ---
    Analytics:
      type: object
      properties:
        period:
          type: string
        contentGenerated:
          type: integer
        campaignsRun:
          type: integer
        keywordsTracked:
          type: integer
        competitorsAnalyzed:
          type: integer
        contentByType:
          type: array
          items:
            $ref: "#/components/schemas/ChartDataPoint"
        campaignActivity:
          type: array
          items:
            $ref: "#/components/schemas/ChartDataPoint"
      required:
        - period
        - contentGenerated
        - campaignsRun
        - keywordsTracked
        - competitorsAnalyzed
        - contentByType
        - campaignActivity
    ChartDataPoint:
      type: object
      properties:
        label:
          type: string
        value:
          type: integer
      required:
        - label
        - value
    ActivityItem:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - type
        - description
        - createdAt

    # --- SEO schemas ---
...