# lib/api-zod/src/generated/types/analytics.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Marketing Agent API
 * OpenAPI spec version: 0.1.0
 */
import type { ChartDataPoint } from './chartDataPoint';

export interface Analytics {
  period: string;
  contentGenerated: number;
  campaignsRun: number;
  keywordsTracked: number;
  competitorsAnalyzed: number;
  contentByType: ChartDataPoint[];
  campaignActivity: ChartDataPoint[];
}