# lib/api-zod/src/generated/types/businessInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Marketing Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface BusinessInput {
  name: string;
  industry: string;
  description: string;
  website?: string;
  targetAudience?: string;
}