# lib/api-zod/src/generated/types/businessStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Marketing Agent API
 * OpenAPI spec version: 0.1.0
 */
import type { BusinessStatsContentByType } from './businessStatsContentByType';

export interface BusinessStats {
  totalContent: number;
  activeCampaigns: number;
  totalKeywords: number;
  totalCompetitors: number;
  contentByType: BusinessStatsContentByType;
}